chore: reorganize imports and add missing __init__.py files for network and routers modules
