+++ src/autodialer/routers/__init__.py
@@ -1,0 +1,4 @@
+from .base_router_api import RouterAPI
+from .get_router import get_router
+
+__all__ = ["RouterAPI", "get_router"]
